Updated project documentation (#30)
