+++ youtube.com/dotNET Microservices by Julio Casal/Playstore/Playstore.Catalog.Contracts/MassTransit/CatalogItemDeleted.cs
@@ -1,0 +1,5 @@
+namespace Playstore.Catalog.Contracts.MassTransit;
+
+/// <summary>Result of catalog item deleted item.</summary>
+/// <param name="ItemId">ID of the item.</param>
+public record CatalogItemDeleted(Guid ItemId);
