--- youtube.com/dotNET Microservices by Julio Casal/Playstore/Playstore.Catalog.Contracts/MassTransit/Contracts.cs
@@ -1,7 +1,0 @@
-namespace Playstore.Catalog.Contracts.MassTransit;
-
-public record CatalogItemCreated(Guid ItemId, string Name, string Description);
-
-public record CatalogItemUpdated(Guid ItemId, string Name, string Description);
-
-public record CatalogItemDeleted(Guid ItemId);
