+++ youtube.com/dotNET Microservices by Julio Casal/Playstore/Playstore.Client/readme.md
@@ -1,0 +1,2 @@
+# Playstore.Client project
+.NET 6 MVC client project that acts as a façade for the microservice functionality.
