+++ youtube.com/dotNET Microservices by Julio Casal/Playstore/Playstore.Inventory.Contracts/readme.md
@@ -1,0 +1,3 @@
+# Playstore.Catalog.Contracts
+
+Inventory contracts shared between projects.
