+++ youtube.com/dotNET Microservices by Julio Casal/Playstore/Playstore.Inventory.Service/Consumers/CatalogItemCreatedConsumer.cs
@@ -5,15 +5,21 @@ using Playstore.Inventory.Service.Entities;
 
 namespace Playstore.Inventory.Service.Consumers;
 
+/// <summary>Consumer of catalog item created messages through MassTransit.</summary>
 public class CatalogItemCreatedConsumer : IConsumer<CatalogItemCreated>
 {
+    /// <summary>Repository to use for syncing data.</summary>
     private readonly IRepository<CatalogItem> _catalogItemRepository;
 
+    /// <summary>Create new class instance.</summary>
+    /// <param name="catalogItemRepository">Repository to use for syncing data.</param>
     public CatalogItemCreatedConsumer(IRepository<CatalogItem> catalogItemRepository)
     {
         _catalogItemRepository = catalogItemRepository;
     }
 
+    /// <summary>Synchronizes local database with the result of the action.</summary>
+    /// <param name="context">Action context.</param>
     public async Task Consume(ConsumeContext<CatalogItemCreated> context)
     {
         CatalogItemCreated message = context.Message;
